requirements.txt
torch>=2.2.0
torchaudio>=2.2.0
demucs>=4.0.0
faster-whisper>=0.10.0
nemo-toolkit>=2.2.1
numpy>=1.24.0
librosa>=0.10.0
soundfile>=0.12.0
scipy>=1.11.0
tqdm>=4.65.0
ctranslate2==4.5.0
git+https://github.com/MahmoudAshraf97/demucs.git
git+https://github.com/oliverguhr/deepmultilingualpunctuation.git
git+https://github.com/MahmoudAshraf97/ctc-forced-aligner.git
omegaconf
nltk
wget 